pass state into to-dos and render each one
Create an UpcomingList to hold not yet to be complete To-Do’s.  Within
UpComingList.js, use
{Object.keys(this.props.todos).map(this.renderToDo)} to iterate through
each todo and pass the key to the renderToDo function. Within
renderToDo, a <ToDo /> component is called and passed the details of
the ToDo through props. These then get rendered in the ToDo.js file as
text.