Build list render logic for complete vs incomplete
Create singular list object that either is false for incomplete, or
true for complete. Within list.js’s renderToDo(key) function, check to
see what the list type is, and then only show ToDo’s that have that
boolean operator. This ensures that one list can be used for incomplete
ToDos and another list can be used for Completed ToDos.