enable adding of todos from submit form
within AddBar.js, wire up the form via refs to allow for collection of
the text within the box. Then on the form itself, make an onSubmit
property call createToDo. CreateToDo will create a toDo using the info
submitted, and then pass this toDo to an addToDo function passed via
props from the App.js level.

In App.js, addToDo(todo) takes in a todo and adds it to the todos
state, and then sets state with the newly created object.