create toggling for done and undone on ToDos
Within ToDo.js, create a label that is set by whether the ToDo is done
(true) or not (false). Set this dynamically generated label as the
content of the <button>. Additionally, make this button TOGGLE the
current button’s complete attribute.

To do so, get rid of the App.js’ completeToDo function and replace it
with a toggleToDo function that gets passed down the prop chain.