allow user to delete todos permanently
within App.js, create a removeToDo function that takes in a key and
then sets it to null (this is to delete from firebase… otherwise you
could simply do “delete todos[key]”… then use this.setState.

Pass removeToDo down the prop chain and set the onClick for the Delete
link to pass in the index of that ToDo to run the removeToDo function.