set up manual routing through address bar
1) In Package.json, need to wind back React because it appears 15.4
isn’t compatible with react-router 4.0 (also used the alpha .3 instead
of .6 latest version).

2) In Index.js, import react router packages and create a stateless
component as Root. This holds the Router and the match cases. At base
of the app, a new ListPicker component will be shown. If a parameter is
given after the /, this will be the name of the list. Then render this
Root component instead of app.

3) Within App.js, change the componentWillMount function to sync
dynamically with wherever the router is parked at the time via the
listId.

4) Add the scaffolding for a listPicker component. Still needs to be
wired up to allow for changing the url to route to a particular list.